# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_server.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  internal_server.renewAccessTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_server.renewAccessTokenResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for the github.com/0xdbb/eggsplore service
  title: github.com/0xdbb/eggsplore API
  version: "1.0"
paths:
...
  /auth/renew:
    post:
      consumes:
      - application/json
      description: Generates a new access token using a valid refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_server.renewAccessTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server.renewAccessTokenResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "401":
          description: Unauthorized or Invalid token
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_server.ErrorResponse'
      summary: Renew Access Token
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"